+++ Changelog.md
@@ -92,3 +92,4 @@ _osl_panic for string "32KHz LPO Clock not running" does not produce panic in Ai
 
 #### v2.0.4
 - Minor improvement in method findService (based on panic report analysis)
+- Do not use removeObject, use getNextObject